Removed RolloutBuffer.py and put flush into mem buffer
