+++ cares_reinforcement_learning/memory/__init__.py
@@ -1,3 +1,2 @@
 from .MemoryBuffer import MemoryBuffer
-from .RolloutBuffer import RolloutBuffer
 from .PrioritizedMemoryBuffer import PrioritizedMemoryBuffer
